fix: update interval column type to double precision to support decimal SRS intervals (e.g. 0.2)

+++ scenario_schema.sql
@@ -36,7 +36,7 @@ CREATE TABLE scenario_vocabulary (
   
   -- SRS Properties
   next_review_date BIGINT NOT NULL,
-  interval INTEGER NOT NULL DEFAULT 0,
+  interval DOUBLE PRECISION NOT NULL DEFAULT 0,
   repetition INTEGER NOT NULL DEFAULT 0,
   ease_factor DOUBLE PRECISION NOT NULL DEFAULT 2.5
 );
